+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.1 under development
 
-- no changes in this release.
+- New #2: Add range request support (@samdark)
 
 ## 1.1.0 April 09, 2026